backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.36
psycopg[binary]
pydantic[email]<2.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
stripe==7.9.0
